Fix Metamask window.web3 injection deprecation
Since Metamask is no longer injecting window.web3 as provider the code
is updated accordingly. We use web3.js as convenience library for
instantiating contracts and interacting with them.

Even though some
methods are similar, most of the code needs refactoring for how to call
the contract methods and for the returned values. Current version of
web3.js (+v1.2) mostly returns String values compared to the outdated
injected web3 (v0.20).

ES6 syntax is used throughout the updated code as users of the dApp are
expected to have Metamask installed, which is available for Chrome and
Firefox. Both browsers support ES6 syntax since around 2014 releases.